+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 ### Fixed
+-   RESTfulRPCAdapter: 
+    -   [Issue #7](https://github.com/repplix/Jexxa/issues/7) Stack trace is now included when serializing an Exception  
 
 ### Changed